+++ kolibri/core/assets/src/core-app/apiSpec.js
@@ -91,7 +91,7 @@ import kCircularLoader from '../views/k-circular-loader';
 
 import multiPaneLayout from '../views/multi-pane-layout';
 import navComponents from '../utils/navComponents';
-import ParseErrors from '../utils/ParseErrors';
+import CatchErrors from '../utils/CatchErrors';
 import * as client from './client';
 import urls from './urls';
 
@@ -187,6 +187,6 @@ export default {
     i18n,
     navComponents,
     samePageCheckGenerator,
-    ParseErrors,
+    CatchErrors,
   },
 };
